refactor(client): move ApiError and HttpError to errors.ts
Classes are values and do not belong in types.ts. No API changes.
